# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend de l'application NoLedge, une plateforme de quiz en ligne",
        "title": "NoLedge API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/proposition/all": {
            "get": {
                "description": "Récupère la liste de toutes les propositions disponibles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proposition"
                ],
                "summary": "Récupérer toutes les propositions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/proposition.Proposition"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/proposition/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crée une nouvelle proposition",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proposition"
                ],
                "summary": "Créer une nouvelle proposition",
                "parameters": [
                    {
                        "description": "Informations de la proposition",
                        "name": "proposition",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/proposition.Proposition"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/proposition.Proposition"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/proposition/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Supprime une proposition",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proposition"
                ],
                "summary": "Supprimer une proposition",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la proposition",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/proposition/get": {
            "get": {
                "description": "Récupère les détails d'une proposition spécifique",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proposition"
                ],
                "summary": "Récupérer une proposition par son ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la proposition",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/proposition.Proposition"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/proposition/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Met à jour les informations d'une proposition existante",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proposition"
                ],
                "summary": "Mettre à jour une proposition",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la proposition",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Nouvelles informations de la proposition",
                        "name": "proposition",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/proposition.Proposition"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/proposition.Proposition"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "proposition.Proposition": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "id_question": {
                    "type": "integer"
                },
                "is_correct": {
                    "type": "boolean"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}